Refactor to aviris_df package
So that our main isn't super cluttered.
Not sure this is the best approach, but
it should be a little easier to deal with.
